test: Config redis for room repository test
