+++ src/main/kotlin/smalltalk/backend/config/redis/RedisConfig.kt
@@ -1,7 +1,5 @@
 package smalltalk.backend.config.redis
 
-import com.fasterxml.jackson.databind.ObjectMapper
-import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory
@@ -9,24 +7,13 @@ import org.springframework.data.redis.core.RedisTemplate
 import org.springframework.data.redis.serializer.StringRedisSerializer
 
 @Configuration
-class RedisConfig(
-    @Value("\${spring.redis.host}")
-    private val host: String,
-    @Value("\${spring.redis.port}")
-    private val port: Int
-) {
+class RedisConfig {
 
     @Bean
-    fun objectMapper() = ObjectMapper()
-
-    @Bean
-    fun redisConnectionFactory() = LettuceConnectionFactory(host, port)
-
-    @Bean
-    fun redisTemplate() =
+    fun redisTemplate(redisConnectionFactory: LettuceConnectionFactory) =
         RedisTemplate<String, Any>().apply {
             keySerializer = StringRedisSerializer()
             valueSerializer = StringRedisSerializer()
-            connectionFactory = redisConnectionFactory()
+            connectionFactory = redisConnectionFactory
         }
 }
