+++ src/test/kotlin/smalltalk/backend/config/redis/TestRedisConfig.kt
@@ -1,0 +1,32 @@
+package smalltalk.backend.config.redis
+
+import com.fasterxml.jackson.databind.ObjectMapper
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.boot.test.context.TestConfiguration
+import org.springframework.context.annotation.Bean
+import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory
+import org.springframework.data.redis.core.RedisTemplate
+import org.springframework.data.redis.serializer.StringRedisSerializer
+
+@TestConfiguration
+class TestRedisConfig(
+    @Value("\${spring.redis.host}")
+    private val host: String,
+    @Value("\${spring.redis.port}")
+    private val port: Int
+) {
+
+    @Bean
+    fun objectMapper() = ObjectMapper()
+
+    @Bean
+    fun redisConnectionFactory() = LettuceConnectionFactory(host, port)
+
+    @Bean
+    fun redisTemplate() =
+        RedisTemplate<String, Any>().apply {
+            keySerializer = StringRedisSerializer()
+            valueSerializer = StringRedisSerializer()
+            connectionFactory = redisConnectionFactory()
+        }
+}
